+++ src/clusterxx/data_structures/kd_tree/kd_tree_impl.hpp
@@ -46,7 +46,7 @@ clusterxx::kd_tree<Metric, PairwiseMetric>::__initialize(
 template <typename Metric, typename PairwiseMetric>
 void clusterxx::kd_tree<Metric, PairwiseMetric>::__k_nearest_neighbors(
     std::unique_ptr<kd_node> &node, const arma::vec &X, MaxHeap &heap,
-    const int depth, const int k) {
+    const int depth, const uint32_t k) {
     if (!node) {
         return;
     }
@@ -132,7 +132,7 @@ void clusterxx::kd_tree<Metric, PairwiseMetric>::__radius_nearest_neighbors(
 }
 
 template <typename Metric, typename PairwiseMetric>
-int clusterxx::kd_tree<Metric, PairwiseMetric>::__depth(
+uint64_t clusterxx::kd_tree<Metric, PairwiseMetric>::__depth(
     std::unique_ptr<kd_node> &root) {
     if (!root) {
         return 0;
@@ -184,7 +184,7 @@ clusterxx::kd_tree<Metric, PairwiseMetric>::query_radius(const arma::vec &X,
 }
 
 template <typename Metric, typename PairwiseMetric>
-int clusterxx::kd_tree<Metric, PairwiseMetric>::depth() {
+uint64_t clusterxx::kd_tree<Metric, PairwiseMetric>::depth() {
     return __depth(__root);
 }
